init routing and create a home route
