+++ src/App.js
@@ -1,35 +1,12 @@
-import Directory from './components/directory/directory.component';
+import { Routes, Route } from 'react-router-dom';
+import Home from './routes/home/home.component';
 
 const App = () => {
-	const categories = [
-		{
-			id: 1,
-			title: 'hats',
-			imageUrl: 'https://i.ibb.co/cvpntL1/hats.png'
-		},
-		{
-			id: 2,
-			title: 'jackets',
-			imageUrl: 'https://i.ibb.co/px2tCc3/jackets.png'
-		},
-		{
-			id: 3,
-			title: 'sneakers',
-			imageUrl: 'https://i.ibb.co/0jqHpnp/sneakers.png'
-		},
-		{
-			id: 4,
-			title: 'womens',
-			imageUrl: 'https://i.ibb.co/GCCdy8t/womens.png'
-		},
-		{
-			id: 5,
-			title: 'mens',
-			imageUrl: 'https://i.ibb.co/R70vBrQ/men.png'
-		}
-	];
-
-	return <Directory categories={categories} />;
+	return (
+		<Routes>
+			<Route path='/home' element={<Home />} />
+		</Routes>
+	);
 };
 
 export default App;
